refactor: put reused codes into 2 components
