+++ packages/extension-polkagate/src/components/ShortAddress2.tsx
@@ -1,13 +1,12 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
-// @ts-nocheck
 
 /* eslint-disable react/jsx-max-props-per-line */
 
+import type { AccountId } from '@polkadot/types/interfaces/runtime';
+
 import { Grid, type SxProps, type Theme } from '@mui/material';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
-
-import type { AccountId } from '@polkadot/types/interfaces/runtime';
 
 import { SHORT_ADDRESS_CHARACTERS } from '../util/constants';
 import CopyAddressButton from './CopyAddressButton';
@@ -21,7 +20,7 @@ interface Props {
   clipped?: boolean;
 }
 
-function ShortAddress2({ address, clipped = false, charsCount = SHORT_ADDRESS_CHARACTERS, style, showCopy = false, inParentheses = false }: Props): React.ReactElement {
+function ShortAddress2 ({ address, charsCount = SHORT_ADDRESS_CHARACTERS, clipped = false, inParentheses = false, showCopy = false, style }: Props): React.ReactElement {
   const [charactersCount, setCharactersCount] = useState<number>(charsCount);
   const pRef = useRef<HTMLDivElement>(null);
   const cRef = useRef<HTMLDivElement>(null);
